@@ -27,21 +27,22 @@
     "flowbite": "^1.5.3",
     "flowbite-react": "^0.2.0",
     "framer-motion": "^7.5.3",
-    "next": "^12.3.1",
+    "lucide-react": "^0.379.0",
+    "next": "14.2.3",
     "next-seo": "^5.5.0",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-code-blocks": "^0.0.9-0",
     "react-dom": "^18.2.0",
     "react-intersection-observer": "^9.4.0",
-    "tabler-icons-react": "^1.55.0",
-    "typewriter-effect": "^2.19.0",
+    "tabler-icons-react": "^1.56.0",
+    "typewriter-effect": "^2.21.0",
     "zustand": "^4.1.2"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.1.2",
     "@commitlint/config-conventional": "^17.1.0",
-    "@next/bundle-analyzer": "^12.3.1",
+    "@next/bundle-analyzer": "14.2.3",
     "@testing-library/cypress": "^8.0.3",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
@@ -56,7 +57,7 @@
     "eslint": "^8.25.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-next": "^12.3.1",
+    "eslint-config-next": "14.2.3",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-import": "^2.26.0",
