@@ -1,9 +1,11 @@
+'use client';
+
 /* eslint-disable tailwindcss/no-custom-classname */
 /* eslint-disable tailwindcss/migration-from-tailwind-2 */
 /* eslint-disable react/no-unescaped-entities */
 import dayjs from 'dayjs';
 import Image from 'next/image';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import type { FC } from 'react';
 import React from 'react';
 
@@ -33,7 +35,7 @@ const BlogCard: FC<BlogCardType> = ({
           {/* image */}
           <div className="h-full w-full ">
             <Image
-              src={featuredImage.url}
+              src={`${process.env.NEXT_PUBLIC_BASE_URL}${featuredImage}`}
               className="rounded-t-lg"
               layout="fill"
               objectFit="cover"
