@@ -15,7 +15,7 @@ const ImageBlock: FC<ImageBlockType> = ({ image }) => {
   return (
     <div className="my-6" key={image.id}>
       <Image
-        src={image.url}
+        src={`${process.env.NEXT_PUBLIC_BASE_URL}${image.url}`}
         height={500}
         width={800}
         alt="notionlink project"
