@@ -1,16 +1,17 @@
-import type { NextPage } from 'next';
+'use client';
+
+import React from 'react';
 
 import { About } from '@/components';
-import { getAllTags } from '@/graphql/Main';
 import { Meta } from '@/layouts/Meta';
 import { Main } from '@/templates/Main';
 import type { TagType } from '@/types/component.types';
-import { axiosGraphQL } from '@/utils/axios';
 
-type AboutPageType = {
+type AboutPageProps = {
   tags: TagType[];
 };
-const AboutPage: NextPage<AboutPageType> = ({ tags }) => {
+
+const AboutPage = ({ tags }: AboutPageProps) => {
   return (
     <Main
       meta={
@@ -23,18 +24,6 @@ const AboutPage: NextPage<AboutPageType> = ({ tags }) => {
       <About tags={tags} />
     </Main>
   );
-};
-
-export const getServerSideProps = async () => {
-  const res = await axiosGraphQL.post(`/`, {
-    query: getAllTags,
-  });
-  const tags = res.data?.data?.Tags.docs;
-  return {
-    props: {
-      tags,
-    },
-  };
 };
 
 export default AboutPage;
